+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Added/Changed/Deprecated/Removed/Fixed/Security: YOUR CHANGE HERE
+* Added: controller action instrumentation [@povilasjurcys](https://github.com/povilasjurcys)
 
 ## 0.7.0 (2019-05-15)